readd height to header datatype

+++ binary_trees.h
@@ -42,6 +42,6 @@ int binary_tree_is_perfect(const binary_tree_t *tree);
 binary_tree_t *binary_tree_sibling(binary_tree_t *node);
 binary_tree_t *binary_tree_uncle(binary_tree_t *node);
 
-size_t binary_tree_height(const binary_tree_t *tree);
+int binary_tree_height(const binary_tree_t *tree);
 void binary_tree_print(const binary_tree_t *);
 #endif /*_BINARY_TREES_H_*/
